--- ExpressionCore.php
@@ -1,18 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: optimistex
- * Date: 23.08.16
- * Time: 13:30
- */
-
-namespace optimistex\expression;
-
-/**
- * Class Expression adapted for Yii2
- * @package optimistex\expression
- */
-class ExpressionCore extends \Expression
-{
-
-}
